csm-1b: bf16 perf win (~1.8x per-frame) + kernel/perf roadmap
Validated bf16 (CsmRMSNorm self-upcasts variance to fp32): backbone cosine
0.999968, argmax 100%, no collapse. Measured: backbone step 48->35ms (1.4x),
depth 291->156ms (1.9x), per-frame compute ~339->191ms. Applied to
generate_speech.py (bf16 model, codec kept fp32 since bf16 breaks Mimi convs;
codec boundary is int codes so dtype-agnostic).

Adds bench_ttft.py (warm per-component latency harness), bench_bf16.py
(bf16-vs-fp32 correctness+speed), and KERNEL_AND_PERF_PLAN.md — the roadmap to
<500ms/<100ms grounded in measured data + the NKI TKG megakernels
(transformer_tkg / attention_block_tkg) as the dispatch-overhead fix, plus
on-device depth loop and TP=2-4.

+++ csm-1b/native-pytorch/src/bench_ttft.py
@@ -66,7 +66,10 @@ def main():
 
     dev = xm.xla_device()
     proc = AutoProcessor.from_pretrained(MODEL)
-    model = CsmForConditionalGeneration.from_pretrained(MODEL, dtype=torch.float32).eval()
+    # Full bf16 — validated clean (cosine 0.999968, argmax 100%, no collapse; CsmRMSNorm
+    # self-upcasts variance to fp32). ~1.4x on the backbone, no dtype boundaries.
+    model = CsmForConditionalGeneration.from_pretrained(MODEL, dtype=torch.bfloat16).eval()
+    model.codec_model = model.codec_model.float()  # codec fp32 (bf16 breaks its convs); fed int codes
     _offload_timed(model.backbone_model, dev, "backbone")
     _offload_timed(model.codec_model, dev, "codec", method="decode")
